- [x] This pull request must be merged before pull request #11.
- [ ] This is a companion pull request to
flatironinstitute/masala_dev_quantum_computing_plugins#11, and should be
merged at the same time as that one.
- [x] This is a companion pull request to
flatironinstitute/masala_dev_standard_plugins#15, and should be merged
at the same time as that one.

Tasks:
- [x] Add a `MasalaFileInterpreterManager`, for managing classes that
permit read or write of particular file types.
- [x] Have it store and index `MasalaFileInterpreter` creator objects.
     - [x] Add a base class for `MasalaFileInterpreterCreator` objects.
     - [x] Add a base class for `MasalaFileInterpreterAPI` objects.
- [x] Update the build scripts to auto-support `MasalaFileInterpreter`
subclasses.
- [x] Update the `MasalaPluginManager` to auto-register plugins that are
file interpreters with the `MasalaFileInterpreterManager`.
     - [x] Unregister, too.

+++ buildme.sh
@@ -4,7 +4,10 @@ mkdir build && echo "Created build/ directory." || echo "Directory build/ alread
 cd build
 mkdir cmake_generated && echo "Created build/cmake_generated/ sub-directory." || echo "Sub-directory build/cmake_generated/ already exists."
 cmake ../cmake/
-make -j `nproc` && echo "Build completed." || echo "Build errors!  Check output log."
+make -j `nproc`
+
+test "${PIPESTATUS[0]}" != '0' && { echo "Build errors!  Check output log for details."; exit 1; } || { echo "Build completed."; }
+
 cd ..
 mkdir headers && echo "Created headers/ directory." || echo "Directory headers/ already exists."
 rm -r headers/*
@@ -13,3 +16,4 @@ python3 ./code_templates/copy_headers.py masala base src headers
 python3 ./code_templates/copy_headers.py masala numeric_api src headers
 python3 ./code_templates/copy_headers.py masala core_api src headers
 python3 ./code_templates/copy_headers.py masala numeric src headers
+echo "Finished building project and copying headers."
